テスト固定
 - 固定ハッシュ、暗号、jsonテスト
