+++ tests/crypto/test_crypto.hpp
@@ -1,8 +1,8 @@
 // Copyright (c) 2012-2018, The CryptoNote developers, The Bytecoin developers, [ ] developers.
-// Licensed under the GNU Lesser General Public License. See LICENSING.md for details.
+// Licensed under the GNU Lesser General Public License. See LICENSE for details.
 
 #pragma once
 
 #include <string>
 
-void test_crypto(const std::string & test_vectors_filename);
+void test_crypto(const std::string &test_vectors_filename);
