+++ tests/hash/test_hash.hpp
@@ -1,8 +1,8 @@
 // Copyright (c) 2012-2018, The CryptoNote developers, The Bytecoin developers, [ ] developers.
-// Licensed under the GNU Lesser General Public License. See LICENSING.md for details.
+// Licensed under the GNU Lesser General Public License. See LICENSE for details.
 
 #pragma once
 
 #include <string>
 
-void test_hashes(const std::string & test_vectors_folder);
+void test_hashes(const std::string &test_vectors_folder);
